requirements.txt
googlesearch-python
pandas
openpyxl
fastapi
uvicorn
jinja2
python-multipart
aiofiles